move DB interface to public types
